Release hotfix JungleJoy.2.2

+++ cg_broker/models.py
@@ -12,6 +12,7 @@ import boto3
 import structlog
 import transip.service
 from suds import WebFault
+from sqlalchemy.types import JSON
 
 import cg_broker
 import cg_sqlalchemy_helpers as cgs
@@ -122,7 +123,7 @@ class Runner(Base, mixins.TimestampMixin, mixins.UUIDMixin):
         'Job', foreign_keys=job_id, back_populates='runners'
     )
 
-    def cleanup_runner(self) -> None:
+    def cleanup_runner(self, shutdown_only: bool) -> None:
         raise NotImplementedError
 
     def start_runner(self) -> None:
@@ -236,7 +237,7 @@ class Runner(Base, mixins.TimestampMixin, mixins.UUIDMixin):
             )
         )
 
-    def kill(self, *, maybe_start_new: bool) -> None:
+    def kill(self, *, maybe_start_new: bool, shutdown_only: bool) -> None:
         """Kill this runner and maybe start a new one.
 
         :param maybe_start_new: Should we maybe start a new runner after
@@ -244,7 +245,7 @@ class Runner(Base, mixins.TimestampMixin, mixins.UUIDMixin):
         """
         self.state = RunnerState.cleaning
         db.session.commit()
-        self.cleanup_runner()
+        self.cleanup_runner(shutdown_only)
         self.state = RunnerState.cleaned
         db.session.commit()
 
@@ -283,7 +284,7 @@ class DevRunner(Runner):
     def start_runner(self) -> None:
         pass
 
-    def cleanup_runner(self) -> None:
+    def cleanup_runner(self, shutdown_only: bool) -> None:
         pass
 
 
@@ -297,7 +298,7 @@ class TransipRunner(Runner):
     def start_runner(self) -> None:
         pass
 
-    def cleanup_runner(self) -> None:
+    def cleanup_runner(self, shutdown_only: bool) -> None:
         username = app.config['_TRANSIP_USERNAME']
         key_file = app.config['_TRANSIP_PRIVATE_KEY_FILE']
         vps_service = transip.service.VpsService(
@@ -393,10 +394,15 @@ class AWSRunner(Runner):
             logger.error('Timed out waiting on AWS instance')
             raise TimeoutError
 
-    def cleanup_runner(self) -> None:
+    def cleanup_runner(self, shutdown_only: bool) -> None:
         assert self.instance_id
         client = boto3.client('ec2')
-        client.terminate_instances(InstanceIds=[self.instance_id])
+        if shutdown_only:
+            client.stop_instances(
+                InstanceIds=[self.instance_id], Hibernate=False
+            )
+        else:
+            client.terminate_instances(InstanceIds=[self.instance_id])
 
     __mapper_args__ = {'polymorphic_identity': RunnerType.aws}
 
@@ -411,6 +417,12 @@ class JobState(enum.IntEnum):
     waiting_for_runner = 1
     started = 2
     finished = 3
+
+    @property
+    def is_finished(self) -> bool:
+        """Is this a finished state.
+        """
+        return self == self.finished
 
 
 class Job(Base, mixins.TimestampMixin, mixins.IdMixin):
@@ -440,6 +452,19 @@ class Job(Base, mixins.TimestampMixin, mixins.IdMixin):
     runners: t.List['Runner'] = db.relationship(
         'Runner', back_populates='job', uselist=True
     )
+
+    _job_metadata: t.Optional[t.Dict[str, object]] = db.Column(
+        'job_metadata', JSON, nullable=True, default={}
+    )
+
+    @property
+    def job_metadata(self) -> t.Dict[str, object]:
+        """Metadata connected to this job.
+        """
+        return self._job_metadata or {}
+
+    def update_metadata(self, new_values: t.Dict[str, object]) -> None:
+        self._job_metadata = {**self.job_metadata, **new_values}
 
     _wanted_runners: int = db.Column(
         'wanted_runners',
@@ -493,4 +518,5 @@ class Job(Base, mixins.TimestampMixin, mixins.IdMixin):
         return {
             'id': self.remote_id,
             'state': self.state.name,
+            'wanted_runners': self.wanted_runners,
         }
